+++ package.json
@@ -15,10 +15,12 @@
     "ES6"
   ],
   "files": [
-    "lib"
+    "lib",
+    "typings.json"
   ],
   "scripts": {
-    "test": "gulp test"
+    "test": "gulp test",
+    "postinstall": "echo \"Nuimo-client-ts is installed.\n Now run 'typings install --ambient noble node' to install the required typings\""
   },
   "dependencies": {
     "noble": "^1.3.0",
@@ -46,5 +48,5 @@
     "node": ">=4.0.0"
   },
   "main": "./lib/index.js",
-  "typings": "./lib/nuimo-client-ts.d.ts"
+  "typings": "./lib/index.d.ts"
 }
