+++ src/noblePeripheral/common-uuids.ts
