+++ src/update/update.ts
@@ -1,6 +1,6 @@
 import "source-map-support/register";
 
-import uuids from "./common-uuids";
+import uuids from "../noblePeripheral/common-uuids";
 
 /**
  * @class
@@ -108,6 +108,3 @@ export function updateFactory(characteristicUUID: string, buffer: Buffer): Updat
       throw new Error(`unidenfied update '${characteristicUUID}'`);
   }
 }
-
-
-export default { updateFactory, SwipeUpdate, ClickUpdate, TurnUpdate, Update };
