+++ tsconfig.json
@@ -16,11 +16,12 @@
   ],
   "files": [
     "typings/main.d.ts",
-    "src/client.ts",
-    "src/clientFactory.ts",
-    "src/common-uuids.ts",
     "src/index.ts",
-    "src/interfaces.d.ts",
-    "src/update.ts"
+    "src/client/clientFactory.ts",
+    "src/client/interfaces.d.ts",
+    "src/client/nuimoClient.ts",
+    "src/noblePeripheral/common-uuids.ts",
+    "src/noblePeripheral/connectToNuimo.ts",
+    "src/update/update.ts"
   ]
 }
